backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai==1.12.0
pydantic
pytz
amadeus
httpx==0.24.1
requests
cachetools


